% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcStrahlerOrder}
\alias{calcStrahlerOrder}
\title{Calculate stream order along network}
\usage{
calcStrahlerOrder(river)
}
\arguments{
\item{river:}{river networks as sf object}
}
\value{
sf object river network, with stream order as an attribute
}
\description{
Given a sarn network, calculate the Strahler stream order. Note that this function was adapted (and simplified) from the hydrostreamer R package implementation of strahler stream order in R (https://github.com/mkkallio/hydrostreamer)
}
