% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{checkCat}
\alias{checkCat}
\title{Network ID column check}
\usage{
checkCat(shapefile)
}
\arguments{
\item{shapefile:}{shapefile to be used in SARNr}
}
\value{
null
}
\description{
Ensures the shapefile ID column is named 'cat'. If it does not exist, create one
}
