% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{checkMask}
\alias{checkMask}
\title{Check river mask format}
\usage{
checkMask(mask)
}
\arguments{
\item{mask:}{raster river mask}
}
\description{
Ensures that river mask is binary composed of only 1s and NaNs
}
