% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{checkRS}
\alias{checkRS}
\title{Check RS band names}
\usage{
checkRS(img)
}
\arguments{
\item{img:}{remote sensing image to be used in SARNr}
}
\description{
Ensures that band names are correct before calculating spectral indices
}
