% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{checkShp}
\alias{checkShp}
\title{Check shapefile object type}
\usage{
checkShp(shapefile)
}
\arguments{
\item{shapefile:}{shapefile to be used in SARNr}
}
\description{
Ensures that shapefiles are sf objects
}
