% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sarn_classifyWater_bimodal}
\alias{sarn_classifyWater_bimodal}
\title{Classifies water}
\usage{
sarn_classifyWater_bimodal(img, dem, maxElev = 4000)
}
\arguments{
\item{img:}{Multi-band RS image to be classified. Required band names are 'green', 'nir', and 'red'}

\item{dem:}{DEM for elevation cutoff of same resolution as RS image}

\item{maxElev:}{Maximum elevation to allow water to exist in meters. Use to avoid ice/snow/glaciers}
}
\description{
Classifies a multi-band image into water/not water using Otsu's bimodal thresholding algorithm + an elevation cutoff.
}
