% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{sarn_data}
\alias{sarn_data}
\title{Produce SARNr object}
\usage{
sarn_data(dem_network, riverMask, dem)
}
\arguments{
\item{dem_network:}{DEM-generated river network as a shapefile}

\item{riverMask:}{River classification raster}

\item{dem:}{Digital elevation model that generated the dem_network}
}
\description{
Produces an SARNr object to use to run the actual algorithm
}
