% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrography.R
\name{sarn_hydrography}
\alias{sarn_hydrography}
\title{Builds actively flowing hydrography from SARN}
\usage{
sarn_hydrography(
  trimmedNetwork,
  dem,
  riverMask,
  lengthThresh,
  printOutput = "Yes"
)
}
\arguments{
\item{trimmedNetwork:}{SARNr-trimmed river network. This would be the $trimmedNetwork output from the sarn_trimNetwork() output}

\item{dem:}{Digital elevation model that was used to generate the DEM river network}

\item{riverMask:}{river mask raster used to generate the RS river network. This would be $rs_raster from the sarn_data() output}

\item{lengthThresh:}{minimum reach length allowed in meters}

\item{printOutput:}{Do you want the results of the iterative trimming to print to console? Default= Yes}
}
\value{
final actively-flowing hydrography
}
\description{
Builds a routing framework and calculates common river network parameters for the river network.
}
