% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_erase}
\alias{st_erase}
\title{Spatial erase function}
\usage{
st_erase(x, y)
}
\arguments{
\item{x:}{layer to erase things away from}

\item{y:}{overlaying layer that defines the 'erase zones'}
}
\description{
Homebaked sf function to mimic the Erase tool in ArcGIS
}
