% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualSARN.R
\name{sarn_joinNetworks}
\alias{sarn_joinNetworks}
\title{SARN network joining}
\usage{
sarn_joinNetworks(data, bufferSize)
}
\arguments{
\item{data:}{sarndata object generated by sarn_data()}

\item{bufferSize:}{Lateral buffer distance for joining river networks (meters)}
}
\value{
sarnnetwork object with two vector river networks: network where DEM and RS are overlapping ('coincident') and not overlapping ('non-coincident')
}
\description{
Joins DEM and RS river networks into a single vector product
}
