% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actualSARN.R
\name{sarn_trimNetworks}
\alias{sarn_trimNetworks}
\title{SARN network clipping}
\usage{
sarn_trimNetworks(networks, crs_code, printOutput = "Yes")
}
\arguments{
\item{networks:}{sarnnetwork object produced by sarn_joinNetworks()}

\item{crs_code:}{epsg code for the projection you are using (for the trimmed nodes)}

\item{printOutput:}{Do you want the results of the iterative trimming to print to console? Default= Yes}
}
\value{
sarnnetwork object with trimmed river network (output$trimmedNetwork)
}
\description{
Iteratively trims river network to the actively-flowing network
}
